Mesh 534/statistic module (#170)
* Moving Asset test

* Statistic test with TestStorage

* Statistic: Test Investors per asset

* Fix style

* Moving Asset test

* Add transfer stats to mint & redeem

* Rename test files

* Fix Style
