+++ runtime/src/asset.rs
@@ -56,7 +56,7 @@
 //! - `custodian_allowance`- Returns the allowance provided to a custodian for a given ticker and token holder
 //! - `total_custody_allowance` - Returns the total allowance approved by the token holder.
 
-use crate::{balances, constants::*, general_tm, identity, percentage_tm, utils};
+use crate::{balances, constants::*, general_tm, identity, percentage_tm, statistics, utils};
 use codec::Encode;
 use core::result::Result as StdResult;
 use currency::*;
@@ -83,6 +83,7 @@ pub trait Trait:
     + balances::Trait
     + identity::Trait
     + pallet_session::Trait
+    + statistics::Trait
 {
     /// The overarching event type.
     type Event: From<Event<Self>> + Into<<Self as frame_system::Trait>::Event>;
@@ -132,11 +133,11 @@ pub struct SecurityToken<U> {
 /// struct to store the signed data
 #[derive(codec::Encode, codec::Decode, Default, Clone, PartialEq, Debug)]
 pub struct SignData<U> {
-    custodian_did: IdentityId,
-    holder_did: IdentityId,
-    ticker: Ticker,
-    value: U,
-    nonce: u16,
+    pub custodian_did: IdentityId,
+    pub holder_did: IdentityId,
+    pub ticker: Ticker,
+    pub value: U,
+    pub nonce: u16,
 }
 
 /// struct to store the ticker registration details
@@ -650,6 +651,7 @@ decl_module! {
             for i in 0..investor_dids.len() {
                 Self::_update_checkpoint(&ticker, investor_dids[i], current_balances[i]);
                 <BalanceOf<T>>::insert((ticker, investor_dids[i]), updated_balances[i]);
+                 <statistics::Module<T>>::update_transfer_stats( &ticker, None, Some(updated_balances[i]), values[i]);
                 Self::deposit_event(RawEvent::Issued(
                     ticker,
                     investor_dids[i],
@@ -706,6 +708,8 @@ decl_module! {
 
             <BalanceOf<T>>::insert((ticker, did), updated_burner_balance);
             <Tokens<T>>::insert(&ticker, token);
+            <statistics::Module<T>>::update_transfer_stats( &ticker, Some(updated_burner_balance), None, value);
+
 
             Self::deposit_event(RawEvent::Redeemed(ticker, did, value));
 
@@ -763,6 +767,7 @@ decl_module! {
             <Allowance<T>>::insert(&ticker_from_did_did, updated_allowance);
             <BalanceOf<T>>::insert(&ticker_did, updated_burner_balance);
             <Tokens<T>>::insert(&ticker, token);
+            <statistics::Module<T>>::update_transfer_stats( &ticker, Some(updated_burner_balance), None, value);
 
             Self::deposit_event(RawEvent::Redeemed(ticker, did, value));
             Self::deposit_event(RawEvent::Approval(ticker, from_did, did, value));
@@ -811,6 +816,7 @@ decl_module! {
 
             <BalanceOf<T>>::insert(&ticker_token_holder_did, updated_burner_balance);
             <Tokens<T>>::insert(&ticker, token);
+            <statistics::Module<T>>::update_transfer_stats( &ticker, Some(updated_burner_balance), None, value);
 
             Self::deposit_event(RawEvent::ControllerRedemption(ticker, did, token_holder_did, value, data, operator_data));
 
@@ -1532,7 +1538,15 @@ impl<T: Trait> Module<T> {
         // increase receiver's balance
         <BalanceOf<T>>::insert(ticker_to_did, updated_to_balance);
 
-        Self::deposit_event(RawEvent::Transfer(*ticker, from_did, to_did, value));
+        // Update statistic info.
+        <statistics::Module<T>>::update_transfer_stats(
+            ticker,
+            Some(updated_from_balance),
+            Some(updated_to_balance),
+            value,
+        );
+
+        Self::deposit_event(RawEvent::Transfer(ticker.clone(), from_did, to_did, value));
         Ok(())
     }
 
@@ -1760,1659 +1774,4 @@ impl<T: Trait> Module<T> {
 
         Ok(())
     }
-}
-
-/// tests for this module
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{exemption, group, identity};
-    use primitives::{IdentityId, Key};
-    use rand::Rng;
-
-    use chrono::prelude::*;
-    use frame_support::{
-        assert_err, assert_noop, assert_ok, dispatch::DispatchResult, impl_outer_origin,
-        parameter_types,
-    };
-    use frame_system::EnsureSignedBy;
-    use lazy_static::lazy_static;
-    use sp_core::{crypto::key_types, H256};
-    use sp_runtime::{
-        testing::{Header, UintAuthorityId},
-        traits::{BlakeTwo256, ConvertInto, IdentityLookup, OpaqueKeys, Verify},
-        AnySignature, KeyTypeId, Perbill,
-    };
-    use std::sync::{Arc, Mutex};
-    use test_client::{self, AccountKeyring};
-
-    type SessionIndex = u32;
-    type AuthorityId = <AnySignature as Verify>::Signer;
-    type BlockNumber = u64;
-    type AccountId = <AnySignature as Verify>::Signer;
-    type OffChainSignature = AnySignature;
-
-    pub struct TestOnSessionEnding;
-    impl pallet_session::OnSessionEnding<AuthorityId> for TestOnSessionEnding {
-        fn on_session_ending(_: SessionIndex, _: SessionIndex) -> Option<Vec<AuthorityId>> {
-            None
-        }
-    }
-
-    pub struct TestSessionHandler;
-    impl pallet_session::SessionHandler<AuthorityId> for TestSessionHandler {
-        const KEY_TYPE_IDS: &'static [KeyTypeId] = &[key_types::DUMMY];
-        fn on_new_session<Ks: OpaqueKeys>(
-            _changed: bool,
-            _validators: &[(AuthorityId, Ks)],
-            _queued_validators: &[(AuthorityId, Ks)],
-        ) {
-        }
-
-        fn on_disabled(_validator_index: usize) {}
-
-        fn on_genesis_session<Ks: OpaqueKeys>(_validators: &[(AuthorityId, Ks)]) {}
-
-        fn on_before_session_ending() {}
-    }
-
-    impl_outer_origin! {
-        pub enum Origin for Test {}
-    }
-
-    // For testing the module, we construct most of a mock runtime. This means
-    // first constructing a configuration type (`Test`) which `impl`s each of the
-    // configuration traits of modules we want to use.
-    #[derive(Clone, Eq, PartialEq)]
-    pub struct Test;
-    parameter_types! {
-        pub const Period: BlockNumber = 1;
-        pub const Offset: BlockNumber = 0;
-        pub const BlockHashCount: u32 = 250;
-        pub const MaximumBlockWeight: u32 = 4 * 1024 * 1024;
-        pub const MaximumBlockLength: u32 = 4 * 1024 * 1024;
-        pub const AvailableBlockRatio: Perbill = Perbill::from_percent(75);
-    }
-    impl frame_system::Trait for Test {
-        type Origin = Origin;
-        type Index = u64;
-        type BlockNumber = u64;
-        type Call = ();
-        type Hash = H256;
-        type Hashing = BlakeTwo256;
-        type AccountId = AccountId;
-        type Lookup = IdentityLookup<Self::AccountId>;
-        type Header = Header;
-        type Event = ();
-        type BlockHashCount = BlockHashCount;
-        type MaximumBlockWeight = MaximumBlockWeight;
-        type MaximumBlockLength = MaximumBlockLength;
-        type AvailableBlockRatio = AvailableBlockRatio;
-        type Version = ();
-        type ModuleToIndex = ();
-    }
-
-    parameter_types! {
-        pub const DisabledValidatorsThreshold: Perbill = Perbill::from_percent(33);
-    }
-
-    impl pallet_session::Trait for Test {
-        type OnSessionEnding = TestOnSessionEnding;
-        type Keys = UintAuthorityId;
-        type ShouldEndSession = pallet_session::PeriodicSessions<Period, Offset>;
-        type SessionHandler = TestSessionHandler;
-        type Event = ();
-        type ValidatorId = AuthorityId;
-        type ValidatorIdOf = ConvertInto;
-        type SelectInitialValidators = ();
-        type DisabledValidatorsThreshold = DisabledValidatorsThreshold;
-    }
-
-    impl pallet_session::historical::Trait for Test {
-        type FullIdentification = ();
-        type FullIdentificationOf = ();
-    }
-
-    parameter_types! {
-        pub const ExistentialDeposit: u64 = 0;
-        pub const TransferFee: u64 = 0;
-        pub const CreationFee: u64 = 0;
-        pub const TransactionBaseFee: u64 = 0;
-        pub const TransactionByteFee: u64 = 0;
-    }
-
-    impl balances::Trait for Test {
-        type Balance = u128;
-        type OnFreeBalanceZero = ();
-        type OnNewAccount = ();
-        type Event = ();
-        type DustRemoval = ();
-        type TransferPayment = ();
-        type ExistentialDeposit = ExistentialDeposit;
-        type TransferFee = TransferFee;
-        type CreationFee = CreationFee;
-        type Identity = crate::identity::Module<Test>;
-    }
-
-    impl general_tm::Trait for Test {
-        type Event = ();
-        type Asset = Module<Test>;
-    }
-
-    parameter_types! {
-        pub const One: AccountId = AccountId::from(AccountKeyring::Dave);
-        pub const Two: AccountId = AccountId::from(AccountKeyring::Dave);
-        pub const Three: AccountId = AccountId::from(AccountKeyring::Dave);
-        pub const Four: AccountId = AccountId::from(AccountKeyring::Dave);
-        pub const Five: AccountId = AccountId::from(AccountKeyring::Dave);
-    }
-
-    impl group::Trait<group::Instance1> for Test {
-        type Event = ();
-        type AddOrigin = EnsureSignedBy<One, AccountId>;
-        type RemoveOrigin = EnsureSignedBy<Two, AccountId>;
-        type SwapOrigin = EnsureSignedBy<Three, AccountId>;
-        type ResetOrigin = EnsureSignedBy<Four, AccountId>;
-        type MembershipInitialized = ();
-        type MembershipChanged = ();
-    }
-
-    impl identity::Trait for Test {
-        type Event = ();
-        type Proposal = Call<Test>;
-        type AcceptTransferTarget = Module<Test>;
-        type AddSignerMultiSigTarget = Test;
-    }
-
-    impl crate::multisig::AddSignerMultiSig for Test {
-        fn accept_multisig_signer(_: Signer, _: u64) -> DispatchResult {
-            unimplemented!()
-        }
-    }
-
-    impl percentage_tm::Trait for Test {
-        type Event = ();
-    }
-
-    impl exemption::Trait for Test {
-        type Event = ();
-        type Asset = Module<Test>;
-    }
-
-    parameter_types! {
-        pub const MinimumPeriod: u64 = 3;
-    }
-
-    impl pallet_timestamp::Trait for Test {
-        type Moment = u64;
-        type OnTimestampSet = ();
-        type MinimumPeriod = MinimumPeriod;
-    }
-
-    impl utils::Trait for Test {
-        type Public = AccountId;
-        type OffChainSignature = OffChainSignature;
-        fn validator_id_to_account_id(
-            v: <Self as pallet_session::Trait>::ValidatorId,
-        ) -> Self::AccountId {
-            v
-        }
-    }
-
-    impl Trait for Test {
-        type Event = ();
-        type Currency = balances::Module<Test>;
-    }
-    type Asset = Module<Test>;
-    type Balances = balances::Module<Test>;
-    type Identity = identity::Module<Test>;
-    type GeneralTM = general_tm::Module<Test>;
-
-    lazy_static! {
-        static ref INVESTOR_MAP_OUTER_LOCK: Arc<Mutex<()>> = Arc::new(Mutex::new(()));
-    }
-
-    /// Build a genesis identity instance owned by account No. 1
-    fn identity_owned_by_alice() -> sp_io::TestExternalities {
-        let mut t = frame_system::GenesisConfig::default()
-            .build_storage::<Test>()
-            .unwrap();
-        identity::GenesisConfig::<Test> {
-            owner: AccountKeyring::Alice.public().into(),
-            did_creation_fee: 250,
-        }
-        .assimilate_storage(&mut t)
-        .unwrap();
-        self::GenesisConfig::<Test> {
-            asset_creation_fee: 0,
-            ticker_registration_fee: 0,
-            ticker_registration_config: TickerRegistrationConfig {
-                max_ticker_length: 8,
-                registration_length: Some(10000),
-            },
-            fee_collector: AccountKeyring::Dave.public().into(),
-        }
-        .assimilate_storage(&mut t)
-        .unwrap();
-        sp_io::TestExternalities::new(t)
-    }
-
-    fn make_account(
-        account_id: &AccountId,
-    ) -> StdResult<(<Test as frame_system::Trait>::Origin, IdentityId), &'static str> {
-        let signed_id = Origin::signed(account_id.clone());
-        Balances::make_free_balance_be(&account_id, 1_000_000);
-        Identity::register_did(signed_id.clone(), vec![]);
-        let did = Identity::get_identity(&Key::try_from(account_id.encode())?).unwrap();
-        Ok((signed_id, did))
-    }
-
-    #[test]
-    fn issuers_can_create_and_rename_tokens() {
-        identity_owned_by_alice().execute_with(|| {
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-            // Raise the owner's base currency balance
-            Balances::make_free_balance_be(&owner_acc, 1_000_000);
-
-            // Expected token entry
-            let token = SecurityToken {
-                name: vec![0x01],
-                owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            assert!(!<identity::DidRecords>::exists(
-                Identity::get_token_did(&ticker).unwrap()
-            ));
-            let identifiers = vec![(IdentifierType::default(), b"undefined".to_vec())];
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            assert_err!(
-                Asset::create_token(
-                    owner_signed.clone(),
-                    owner_did,
-                    token.name.clone(),
-                    ticker,
-                    1_000_000_000_000_000_000_000_000, // Total supply over the limit
-                    true,
-                    token.asset_type.clone(),
-                    identifiers.clone(),
-                ),
-                "Total supply above the limit"
-            );
-
-            // Issuance is successful
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token.name.clone(),
-                ticker,
-                token.total_supply,
-                true,
-                token.asset_type.clone(),
-                identifiers.clone(),
-            ));
-
-            // A correct entry is added
-            assert_eq!(Asset::token_details(ticker), token);
-            //assert!(Identity::is_existing_identity(Identity::get_token_did(&token.name).unwrap()));
-            assert!(<identity::DidRecords>::exists(
-                Identity::get_token_did(&ticker).unwrap()
-            ));
-            assert_eq!(Asset::token_details(ticker), token);
-
-            // Unauthorized identities cannot rename the token.
-            let eve_acc = AccountId::from(AccountKeyring::Eve);
-            let (eve_signed, _eve_did) = make_account(&eve_acc).unwrap();
-            assert_err!(
-                Asset::rename_token(eve_signed, ticker, vec![0xde, 0xad, 0xbe, 0xef]),
-                "sender must be a signing key for the token owner DID"
-            );
-            // The token should remain unchanged in storage.
-            assert_eq!(Asset::token_details(ticker), token);
-            // Rename the token and check storage has been updated.
-            let renamed_token = SecurityToken {
-                name: vec![0x42],
-                owner_did: token.owner_did,
-                total_supply: token.total_supply,
-                divisible: token.divisible,
-                asset_type: token.asset_type.clone(),
-            };
-            assert_ok!(Asset::rename_token(
-                owner_signed.clone(),
-                ticker,
-                renamed_token.name.clone()
-            ));
-            assert_eq!(Asset::token_details(ticker), renamed_token);
-            for (typ, val) in identifiers {
-                assert_eq!(Asset::identifiers((ticker, typ)), val);
-            }
-        });
-    }
-
-    /// # TODO
-    /// It should be re-enable once issuer claim is re-enabled.
-    #[test]
-    #[ignore]
-    fn non_issuers_cant_create_tokens() {
-        identity_owned_by_alice().execute_with(|| {
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (_, owner_did) = make_account(&owner_acc).unwrap();
-
-            // Expected token entry
-            let _ = SecurityToken {
-                name: vec![0x01],
-                owner_did: owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-
-            let wrong_acc = AccountId::from(AccountKeyring::Bob);
-
-            Balances::make_free_balance_be(&wrong_acc, 1_000_000);
-
-            let wrong_did = IdentityId::try_from("did:poly:wrong");
-            assert!(wrong_did.is_err());
-        });
-    }
-
-    #[test]
-    fn valid_transfers_pass() {
-        identity_owned_by_alice().execute_with(|| {
-            let now = Utc::now();
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-            // Expected token entry
-            let token = SecurityToken {
-                name: vec![0x01],
-                owner_did: owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            Balances::make_free_balance_be(&owner_acc, 1_000_000);
-
-            let alice_acc = AccountId::from(AccountKeyring::Alice);
-            let (_, alice_did) = make_account(&alice_acc).unwrap();
-
-            Balances::make_free_balance_be(&alice_acc, 1_000_000);
-
-            // Issuance is successful
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token.name.clone(),
-                ticker,
-                token.total_supply,
-                true,
-                token.asset_type.clone(),
-                vec![],
-            ));
-
-            // A correct entry is added
-            assert_eq!(Asset::token_details(ticker), token);
-
-            let asset_rule = general_tm::AssetRule {
-                sender_rules: vec![],
-                receiver_rules: vec![],
-            };
-
-            // Allow all transfers
-            assert_ok!(GeneralTM::add_active_rule(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                asset_rule
-            ));
-
-            assert_ok!(Asset::transfer(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                alice_did,
-                500
-            ));
-        })
-    }
-
-    #[test]
-    fn valid_custodian_allowance() {
-        identity_owned_by_alice().execute_with(|| {
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-            let now = Utc::now();
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-            // Expected token entry
-            let token = SecurityToken {
-                name: vec![0x01],
-                owner_did: owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            Balances::make_free_balance_be(&owner_acc, 1_000_000);
-
-            let investor1_acc = AccountId::from(AccountKeyring::Bob);
-            let (investor1_signed, investor1_did) = make_account(&investor1_acc).unwrap();
-
-            Balances::make_free_balance_be(&investor1_acc, 1_000_000);
-
-            let investor2_acc = AccountId::from(AccountKeyring::Charlie);
-            let (investor2_signed, investor2_did) = make_account(&investor2_acc).unwrap();
-
-            Balances::make_free_balance_be(&investor2_acc, 1_000_000);
-
-            let custodian_acc = AccountId::from(AccountKeyring::Eve);
-            let (custodian_signed, custodian_did) = make_account(&custodian_acc).unwrap();
-
-            Balances::make_free_balance_be(&custodian_acc, 1_000_000);
-
-            // Issuance is successful
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token.name.clone(),
-                ticker,
-                token.total_supply,
-                true,
-                token.asset_type.clone(),
-                vec![],
-            ));
-
-            assert_eq!(
-                Asset::balance_of((ticker, token.owner_did)),
-                token.total_supply
-            );
-
-            assert_eq!(Asset::token_details(ticker), token);
-
-            let asset_rule = general_tm::AssetRule {
-                sender_rules: vec![],
-                receiver_rules: vec![],
-            };
-
-            // Allow all transfers
-            assert_ok!(GeneralTM::add_active_rule(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                asset_rule
-            ));
-            let funding_round1 = b"Round One".to_vec();
-            assert_ok!(Asset::set_funding_round(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                funding_round1.clone()
-            ));
-            // Mint some tokens to investor1
-            let num_tokens1: u128 = 2_000_000;
-            assert_ok!(Asset::issue(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                investor1_did,
-                num_tokens1,
-                vec![0x0]
-            ));
-            assert_eq!(Asset::funding_round(&ticker), funding_round1.clone());
-            assert_eq!(
-                Asset::issued_in_funding_round((ticker, funding_round1.clone())),
-                num_tokens1
-            );
-            // Check the expected default behaviour of the map.
-            assert_eq!(
-                Asset::issued_in_funding_round((ticker, b"No such round".to_vec())),
-                0
-            );
-            assert_eq!(Asset::balance_of((ticker, investor1_did)), num_tokens1,);
-
-            // Failed to add custodian because of insufficient balance
-            assert_noop!(
-                Asset::increase_custody_allowance(
-                    investor1_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    custodian_did,
-                    250_00_00 as u128
-                ),
-                "Insufficient balance of holder did"
-            );
-
-            // Failed to add/increase the custodian allowance because of Invalid custodian did
-            let custodian_did_not_register = IdentityId::from(5u128);
-            assert_noop!(
-                Asset::increase_custody_allowance(
-                    investor1_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    custodian_did_not_register,
-                    50_00_00 as u128
-                ),
-                "Invalid custodian DID"
-            );
-
-            // Add custodian
-            assert_ok!(Asset::increase_custody_allowance(
-                investor1_signed.clone(),
-                ticker,
-                investor1_did,
-                custodian_did,
-                50_00_00 as u128
-            ));
-
-            assert_eq!(
-                Asset::custodian_allowance((ticker, investor1_did, custodian_did)),
-                50_00_00 as u128
-            );
-
-            assert_eq!(
-                Asset::total_custody_allowance((ticker, investor1_did)),
-                50_00_00 as u128
-            );
-
-            // Transfer the token upto the limit
-            assert_ok!(Asset::transfer(
-                investor1_signed.clone(),
-                investor1_did,
-                ticker,
-                investor2_did,
-                140_00_00 as u128
-            ));
-
-            assert_eq!(
-                Asset::balance_of((ticker, investor2_did)),
-                140_00_00 as u128
-            );
-
-            // Try to Transfer the tokens beyond the limit
-            assert_noop!(
-                Asset::transfer(
-                    investor1_signed.clone(),
-                    investor1_did,
-                    ticker,
-                    investor2_did,
-                    50_00_00 as u128
-                ),
-                "Insufficient balance for transfer"
-            );
-
-            // Should fail to transfer the token by the custodian because of invalid signing key
-            assert_noop!(
-                Asset::transfer_by_custodian(
-                    investor2_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    custodian_did,
-                    investor2_did,
-                    45_00_00 as u128
-                ),
-                "sender must be a signing key for DID"
-            );
-
-            // Should fail to transfer the token by the custodian because of insufficient allowance
-            assert_noop!(
-                Asset::transfer_by_custodian(
-                    custodian_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    custodian_did,
-                    investor2_did,
-                    55_00_00 as u128
-                ),
-                "Insufficient allowance"
-            );
-
-            // Successfully transfer by the custodian
-            assert_ok!(Asset::transfer_by_custodian(
-                custodian_signed.clone(),
-                ticker,
-                investor1_did,
-                custodian_did,
-                investor2_did,
-                45_00_00 as u128
-            ));
-        });
-    }
-
-    #[test]
-    fn valid_custodian_allowance_of() {
-        identity_owned_by_alice().execute_with(|| {
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-            let now = Utc::now();
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-            // Expected token entry
-            let token = SecurityToken {
-                name: vec![0x01],
-                owner_did: owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            Balances::make_free_balance_be(&owner_acc, 1_000_000);
-
-            let investor1_acc = AccountId::from(AccountKeyring::Bob);
-            let (investor1_signed, investor1_did) = make_account(&investor1_acc).unwrap();
-
-            Balances::make_free_balance_be(&investor1_acc, 1_000_000);
-
-            let investor2_acc = AccountId::from(AccountKeyring::Charlie);
-            let (investor2_signed, investor2_did) = make_account(&investor2_acc).unwrap();
-
-            Balances::make_free_balance_be(&investor2_acc, 1_000_000);
-
-            let custodian_acc = AccountId::from(AccountKeyring::Eve);
-            let (custodian_signed, custodian_did) = make_account(&custodian_acc).unwrap();
-
-            Balances::make_free_balance_be(&custodian_acc, 1_000_000);
-
-            // Issuance is successful
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token.name.clone(),
-                ticker,
-                token.total_supply,
-                true,
-                token.asset_type.clone(),
-                vec![],
-            ));
-
-            assert_eq!(
-                Asset::balance_of((ticker, token.owner_did)),
-                token.total_supply
-            );
-
-            assert_eq!(Asset::token_details(ticker), token);
-
-            let asset_rule = general_tm::AssetRule {
-                sender_rules: vec![],
-                receiver_rules: vec![],
-            };
-
-            // Allow all transfers
-            assert_ok!(GeneralTM::add_active_rule(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                asset_rule
-            ));
-
-            // Mint some tokens to investor1
-            assert_ok!(Asset::issue(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                investor1_did,
-                200_00_00 as u128,
-                vec![0x0]
-            ));
-
-            assert_eq!(
-                Asset::balance_of((ticker, investor1_did)),
-                200_00_00 as u128
-            );
-
-            let msg = SignData {
-                custodian_did: custodian_did,
-                holder_did: investor1_did,
-                ticker,
-                value: 50_00_00 as u128,
-                nonce: 1,
-            };
-
-            let investor1_key = AccountKeyring::Bob;
-
-            // Add custodian
-            assert_ok!(Asset::increase_custody_allowance_of(
-                investor2_signed.clone(),
-                ticker,
-                investor1_did,
-                investor1_acc.clone(),
-                custodian_did,
-                investor2_did,
-                50_00_00 as u128,
-                1,
-                OffChainSignature::from(investor1_key.sign(&msg.encode()))
-            ));
-
-            assert_eq!(
-                Asset::custodian_allowance((ticker, investor1_did, custodian_did)),
-                50_00_00 as u128
-            );
-
-            assert_eq!(
-                Asset::total_custody_allowance((ticker, investor1_did)),
-                50_00_00 as u128
-            );
-
-            // use the same signature with the same nonce should fail
-            assert_noop!(
-                Asset::increase_custody_allowance_of(
-                    investor2_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    investor1_acc.clone(),
-                    custodian_did,
-                    investor2_did,
-                    50_00_00 as u128,
-                    1,
-                    OffChainSignature::from(investor1_key.sign(&msg.encode()))
-                ),
-                "Signature already used"
-            );
-
-            // use the same signature with the different nonce should fail
-            assert_noop!(
-                Asset::increase_custody_allowance_of(
-                    investor2_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    investor1_acc.clone(),
-                    custodian_did,
-                    investor2_did,
-                    50_00_00 as u128,
-                    3,
-                    OffChainSignature::from(investor1_key.sign(&msg.encode()))
-                ),
-                "Invalid signature"
-            );
-
-            // Transfer the token upto the limit
-            assert_ok!(Asset::transfer(
-                investor1_signed.clone(),
-                investor1_did,
-                ticker,
-                investor2_did,
-                140_00_00 as u128
-            ));
-
-            assert_eq!(
-                Asset::balance_of((ticker, investor2_did)),
-                140_00_00 as u128
-            );
-
-            // Try to Transfer the tokens beyond the limit
-            assert_noop!(
-                Asset::transfer(
-                    investor1_signed.clone(),
-                    investor1_did,
-                    ticker,
-                    investor2_did,
-                    50_00_00 as u128
-                ),
-                "Insufficient balance for transfer"
-            );
-
-            // Should fail to transfer the token by the custodian because of invalid signing key
-            assert_noop!(
-                Asset::transfer_by_custodian(
-                    investor2_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    custodian_did,
-                    investor2_did,
-                    45_00_00 as u128
-                ),
-                "sender must be a signing key for DID"
-            );
-
-            // Should fail to transfer the token by the custodian because of insufficient allowance
-            assert_noop!(
-                Asset::transfer_by_custodian(
-                    custodian_signed.clone(),
-                    ticker,
-                    investor1_did,
-                    custodian_did,
-                    investor2_did,
-                    55_00_00 as u128
-                ),
-                "Insufficient allowance"
-            );
-
-            // Successfully transfer by the custodian
-            assert_ok!(Asset::transfer_by_custodian(
-                custodian_signed.clone(),
-                ticker,
-                investor1_did,
-                custodian_did,
-                investor2_did,
-                45_00_00 as u128
-            ));
-        });
-    }
-
-    #[test]
-    fn checkpoints_fuzz_test() {
-        println!("Starting");
-        for _ in 0..10 {
-            // When fuzzing in local, feel free to bump this number to add more fuzz runs.
-            identity_owned_by_alice().execute_with(|| {
-                let now = Utc::now();
-                <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-                let owner_acc = AccountId::from(AccountKeyring::Dave);
-                let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-                // Expected token entry
-                let token = SecurityToken {
-                    name: vec![0x01],
-                    owner_did: owner_did,
-                    total_supply: 1_000_000,
-                    divisible: true,
-                    asset_type: AssetType::default(),
-                };
-                let ticker = Ticker::from_slice(token.name.as_slice());
-                let bob_acc = AccountId::from(AccountKeyring::Bob);
-                let (_, bob_did) = make_account(&bob_acc).unwrap();
-
-                // Issuance is successful
-                assert_ok!(Asset::create_token(
-                    owner_signed.clone(),
-                    owner_did,
-                    token.name.clone(),
-                    ticker,
-                    token.total_supply,
-                    true,
-                    token.asset_type.clone(),
-                    vec![],
-                ));
-
-                let asset_rule = general_tm::AssetRule {
-                    sender_rules: vec![],
-                    receiver_rules: vec![],
-                };
-
-                // Allow all transfers
-                assert_ok!(GeneralTM::add_active_rule(
-                    owner_signed.clone(),
-                    owner_did,
-                    ticker,
-                    asset_rule
-                ));
-
-                let mut owner_balance: [u128; 100] = [1_000_000; 100];
-                let mut bob_balance: [u128; 100] = [0; 100];
-                let mut rng = rand::thread_rng();
-                for j in 1..100 {
-                    let transfers = rng.gen_range(0, 10);
-                    owner_balance[j] = owner_balance[j - 1];
-                    bob_balance[j] = bob_balance[j - 1];
-                    for _k in 0..transfers {
-                        if j == 1 {
-                            owner_balance[0] -= 1;
-                            bob_balance[0] += 1;
-                        }
-                        owner_balance[j] -= 1;
-                        bob_balance[j] += 1;
-                        assert_ok!(Asset::transfer(
-                            owner_signed.clone(),
-                            owner_did,
-                            ticker,
-                            bob_did,
-                            1
-                        ));
-                    }
-                    assert_ok!(Asset::create_checkpoint(
-                        owner_signed.clone(),
-                        owner_did,
-                        ticker,
-                    ));
-                    let x: u64 = u64::try_from(j).unwrap();
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, owner_did, 0),
-                        owner_balance[j]
-                    );
-                    assert_eq!(Asset::get_balance_at(ticker, bob_did, 0), bob_balance[j]);
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, owner_did, 1),
-                        owner_balance[1]
-                    );
-                    assert_eq!(Asset::get_balance_at(ticker, bob_did, 1), bob_balance[1]);
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, owner_did, x - 1),
-                        owner_balance[j - 1]
-                    );
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, bob_did, x - 1),
-                        bob_balance[j - 1]
-                    );
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, owner_did, x),
-                        owner_balance[j]
-                    );
-                    assert_eq!(Asset::get_balance_at(ticker, bob_did, x), bob_balance[j]);
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, owner_did, x + 1),
-                        owner_balance[j]
-                    );
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, bob_did, x + 1),
-                        bob_balance[j]
-                    );
-                    assert_eq!(
-                        Asset::get_balance_at(ticker, owner_did, 1000),
-                        owner_balance[j]
-                    );
-                    assert_eq!(Asset::get_balance_at(ticker, bob_did, 1000), bob_balance[j]);
-                }
-            });
-        }
-    }
-
-    #[test]
-    fn register_ticker() {
-        identity_owned_by_alice().execute_with(|| {
-            let now = Utc::now();
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-            Balances::make_free_balance_be(&owner_acc, 1_000_000);
-
-            let token = SecurityToken {
-                name: vec![0x01],
-                owner_did: owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-            let identifiers = vec![(IdentifierType::Custom(b"check".to_vec()), b"me".to_vec())];
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            // Issuance is successful
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token.name.clone(),
-                ticker,
-                token.total_supply,
-                true,
-                token.asset_type.clone(),
-                identifiers.clone(),
-            ));
-
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, owner_did), true);
-            assert_eq!(Asset::is_ticker_available(&ticker), false);
-            let stored_token = <Module<Test>>::token_details(&ticker);
-            assert_eq!(stored_token.asset_type, token.asset_type);
-            for (typ, val) in identifiers {
-                assert_eq!(Asset::identifiers((ticker, typ)), val);
-            }
-
-            assert_err!(
-                Asset::register_ticker(owner_signed.clone(), Ticker::from_slice(&[0x01])),
-                "token already created"
-            );
-
-            assert_err!(
-                Asset::register_ticker(
-                    owner_signed.clone(),
-                    Ticker::from_slice(&[0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01])
-                ),
-                "ticker length over the limit"
-            );
-
-            let ticker = Ticker::from_slice(&[0x01, 0x01]);
-
-            assert_eq!(Asset::is_ticker_available(&ticker), true);
-
-            assert_ok!(Asset::register_ticker(owner_signed.clone(), ticker));
-
-            let alice_acc = AccountId::from(AccountKeyring::Alice);
-            let (alice_signed, _) = make_account(&alice_acc).unwrap();
-
-            Balances::make_free_balance_be(&alice_acc, 1_000_000);
-
-            assert_err!(
-                Asset::register_ticker(alice_signed.clone(), ticker),
-                "ticker registered to someone else"
-            );
-
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, owner_did), true);
-            assert_eq!(Asset::is_ticker_available(&ticker), false);
-
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64 + 10001);
-
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, owner_did), false);
-            assert_eq!(Asset::is_ticker_available(&ticker), true);
-        })
-    }
-
-    #[test]
-    fn transfer_ticker() {
-        identity_owned_by_alice().execute_with(|| {
-            let now = Utc::now();
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-            let alice_acc = AccountId::from(AccountKeyring::Alice);
-            let (alice_signed, alice_did) = make_account(&alice_acc).unwrap();
-
-            let bob_acc = AccountId::from(AccountKeyring::Bob);
-            let (bob_signed, bob_did) = make_account(&bob_acc).unwrap();
-
-            let ticker = Ticker::from_slice(&[0x01, 0x01]);
-
-            assert_eq!(Asset::is_ticker_available(&ticker), true);
-            assert_ok!(Asset::register_ticker(owner_signed.clone(), ticker));
-
-            Identity::add_auth(
-                Signer::from(owner_did),
-                Signer::from(alice_did),
-                AuthorizationData::TransferTicker(ticker),
-                None,
-            );
-
-            Identity::add_auth(
-                Signer::from(owner_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTicker(ticker),
-                None,
-            );
-
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, owner_did), true);
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, alice_did), false);
-            assert_eq!(Asset::is_ticker_available(&ticker), false);
-
-            let mut auth_id = Identity::last_authorization(Signer::from(alice_did));
-
-            assert_err!(
-                Asset::accept_ticker_transfer(alice_signed.clone(), auth_id + 1),
-                "Authorization does not exist"
-            );
-
-            assert_ok!(Asset::accept_ticker_transfer(alice_signed.clone(), auth_id));
-
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_ticker_transfer(bob_signed.clone(), auth_id),
-                "Illegal use of Authorization"
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTicker(ticker),
-                Some(now.timestamp() as u64 - 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_ticker_transfer(bob_signed.clone(), auth_id),
-                "Authorization expired"
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::Custom(ticker),
-                Some(now.timestamp() as u64 + 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_ticker_transfer(bob_signed.clone(), auth_id),
-                Error::<Test>::NoTickerTransferAuth
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTicker(ticker),
-                Some(now.timestamp() as u64 + 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_ok!(Asset::accept_ticker_transfer(bob_signed.clone(), auth_id));
-
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, owner_did), false);
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, alice_did), false);
-            assert_eq!(Asset::is_ticker_registry_valid(&ticker, bob_did), true);
-            assert_eq!(Asset::is_ticker_available(&ticker), false);
-        })
-    }
-
-    #[test]
-    fn transfer_token_ownership() {
-        identity_owned_by_alice().execute_with(|| {
-            let now = Utc::now();
-            <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-
-            let alice_acc = AccountId::from(AccountKeyring::Alice);
-            let (alice_signed, alice_did) = make_account(&alice_acc).unwrap();
-
-            let bob_acc = AccountId::from(AccountKeyring::Bob);
-            let (bob_signed, bob_did) = make_account(&bob_acc).unwrap();
-
-            let token_name = vec![0x01, 0x01];
-            let ticker = Ticker::from_slice(token_name.as_slice());
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token_name.clone(),
-                ticker,
-                1_000_000,
-                true,
-                AssetType::default(),
-                vec![],
-            ));
-
-            Identity::add_auth(
-                Signer::from(owner_did),
-                Signer::from(alice_did),
-                AuthorizationData::TransferTokenOwnership(ticker),
-                None,
-            );
-
-            Identity::add_auth(
-                Signer::from(owner_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTokenOwnership(ticker),
-                None,
-            );
-
-            assert_eq!(Asset::token_details(&ticker).owner_did, owner_did);
-
-            let mut auth_id = Identity::last_authorization(Signer::from(alice_did));
-
-            assert_err!(
-                Asset::accept_token_ownership_transfer(alice_signed.clone(), auth_id + 1),
-                "Authorization does not exist"
-            );
-
-            assert_ok!(Asset::accept_token_ownership_transfer(
-                alice_signed.clone(),
-                auth_id
-            ));
-            assert_eq!(Asset::token_details(&ticker).owner_did, alice_did);
-
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_token_ownership_transfer(bob_signed.clone(), auth_id),
-                "Illegal use of Authorization"
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTokenOwnership(ticker),
-                Some(now.timestamp() as u64 - 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_token_ownership_transfer(bob_signed.clone(), auth_id),
-                "Authorization expired"
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::Custom(ticker),
-                Some(now.timestamp() as u64 + 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_token_ownership_transfer(bob_signed.clone(), auth_id),
-                Error::<Test>::NotTickerOwnershipTransferAuth
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTokenOwnership(Ticker::from_slice(&[0x50])),
-                Some(now.timestamp() as u64 + 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_err!(
-                Asset::accept_token_ownership_transfer(bob_signed.clone(), auth_id),
-                "Token does not exist"
-            );
-
-            Identity::add_auth(
-                Signer::from(alice_did),
-                Signer::from(bob_did),
-                AuthorizationData::TransferTokenOwnership(ticker),
-                Some(now.timestamp() as u64 + 100),
-            );
-            auth_id = Identity::last_authorization(Signer::from(bob_did));
-            assert_ok!(Asset::accept_token_ownership_transfer(
-                bob_signed.clone(),
-                auth_id
-            ));
-            assert_eq!(Asset::token_details(&ticker).owner_did, bob_did);
-        })
-    }
-
-    #[test]
-    fn update_identifiers() {
-        identity_owned_by_alice().execute_with(|| {
-            let owner_acc = AccountId::from(AccountKeyring::Dave);
-            let (owner_signed, owner_did) = make_account(&owner_acc).unwrap();
-            // Raise the owner's base currency balance
-            Balances::make_free_balance_be(&owner_acc, 1_000_000);
-            // Expected token entry
-            let token = SecurityToken {
-                name: b"TEST".to_vec(),
-                owner_did,
-                total_supply: 1_000_000,
-                divisible: true,
-                asset_type: AssetType::default(),
-            };
-            let ticker = Ticker::from_slice(token.name.as_slice());
-            assert!(!<identity::DidRecords>::exists(
-                Identity::get_token_did(&ticker).unwrap()
-            ));
-            let identifier_value1 = b"ABC123";
-            let identifiers = vec![(IdentifierType::Cusip, identifier_value1.to_vec())];
-            assert_ok!(Asset::create_token(
-                owner_signed.clone(),
-                owner_did,
-                token.name.clone(),
-                ticker,
-                token.total_supply,
-                true,
-                token.asset_type.clone(),
-                identifiers.clone(),
-            ));
-            // A correct entry was added
-            assert_eq!(Asset::token_details(ticker), token);
-            assert_eq!(
-                Asset::identifiers((ticker, IdentifierType::Cusip)),
-                identifier_value1.to_vec()
-            );
-            let identifier_value2 = b"XYZ555";
-            let updated_identifiers = vec![
-                (IdentifierType::Cusip, Default::default()),
-                (IdentifierType::Isin, identifier_value2.to_vec()),
-            ];
-            assert_ok!(Asset::update_identifiers(
-                owner_signed.clone(),
-                owner_did,
-                ticker,
-                updated_identifiers.clone(),
-            ));
-            for (typ, val) in updated_identifiers {
-                assert_eq!(Asset::identifiers((ticker, typ)), val);
-            }
-        });
-    }
-
-    /*
-     *    #[test]
-     *    /// This test loads up a YAML of testcases and checks each of them
-     *    fn transfer_scenarios_external() {
-     *        let mut yaml_path_buf = PathBuf::new();
-     *        yaml_path_buf.push(env!("CARGO_MANIFEST_DIR")); // This package's root
-     *        yaml_path_buf.push("tests/asset_transfers.yml");
-     *
-     *        println!("Loading YAML from {:?}", yaml_path_buf);
-     *
-     *        let yaml_string = read_to_string(yaml_path_buf.as_path())
-     *            .expect("Could not load the YAML file to a string");
-     *
-     *        // Parse the YAML
-     *        let yaml = YamlLoader::load_from_str(&yaml_string).expect("Could not parse the YAML file");
-     *
-     *        let yaml = &yaml[0];
-     *
-     *        let now = Utc::now();
-     *
-     *        for case in yaml["test_cases"]
-     *            .as_vec()
-     *            .expect("Could not reach test_cases")
-     *        {
-     *            println!("Case: {:#?}", case);
-     *
-     *            let accounts = case["named_accounts"]
-     *                .as_hash()
-     *                .expect("Could not view named_accounts as a hashmap");
-     *
-     *            let mut externalities = if let Some(identity_owner) =
-     *                accounts.get(&Yaml::String("identity-owner".to_owned()))
-     *            {
-     *                identity_owned_by(
-     *                    identity_owner["id"]
-     *                        .as_i64()
-     *                        .expect("Could not get identity owner's ID") as u64,
-     *                )
-     *            } else {
-     *                frame_system::GenesisConfig::default()
-     *                    .build_storage()
-     *                    .unwrap()
-     *                    .0
-     *                    .into()
-     *            };
-     *
-     *            with_externalities(&mut externalities, || {
-     *                // Instantiate accounts
-     *                for (name, account) in accounts {
-     *                    <pallet_timestamp::Module<Test>>::set_timestamp(now.timestamp() as u64);
-     *                    let name = name
-     *                        .as_str()
-     *                        .expect("Could not take named_accounts key as string");
-     *                    let id = account["id"].as_i64().expect("id is not a number") as u64;
-     *                    let balance = account["balance"]
-     *                        .as_i64()
-     *                        .expect("balance is not a number");
-     *
-     *                    println!("Preparing account {}", name);
-     *
-     *                    Balances::make_free_balance_be(&id, balance.clone() as u128);
-     *                    println!("{}: gets {} initial balance", name, balance);
-     *                    if account["issuer"]
-     *                        .as_bool()
-     *                        .expect("Could not check if account is an issuer")
-     *                    {
-     *                        assert_ok!(identity::Module::<Test>::do_create_issuer(id));
-     *                        println!("{}: becomes issuer", name);
-     *                    }
-     *                    if account["investor"]
-     *                        .as_bool()
-     *                        .expect("Could not check if account is an investor")
-     *                    {
-     *                        assert_ok!(identity::Module::<Test>::do_create_investor(id));
-     *                        println!("{}: becomes investor", name);
-     *                    }
-     *                }
-     *
-     *                // Issue tokens
-     *                let tokens = case["tokens"]
-     *                    .as_hash()
-     *                    .expect("Could not view tokens as a hashmap");
-     *
-     *                for (ticker, token) in tokens {
-     *                    let ticker = ticker.as_str().expect("Can't parse ticker as string");
-     *                    println!("Preparing token {}:", ticker);
-     *
-     *                    let owner = token["owner"]
-     *                        .as_str()
-     *                        .expect("Can't parse owner as string");
-     *
-     *                    let owner_id = accounts
-     *                        .get(&Yaml::String(owner.to_owned()))
-     *                        .expect("Can't get owner record")["id"]
-     *                        .as_i64()
-     *                        .expect("Can't parse owner id as i64")
-     *                        as u64;
-     *                    let total_supply = token["total_supply"]
-     *                        .as_i64()
-     *                        .expect("Can't parse the total supply as i64")
-     *                        as u128;
-     *
-     *                    let token_struct = SecurityToken {
-     *                        name: *ticker.into_bytes(),
-     *                        owner: owner_id,
-     *                        total_supply,
-     *                        divisible: true,
-     *                    };
-     *                    println!("{:#?}", token_struct);
-     *
-     *                    // Check that issuing succeeds/fails as expected
-     *                    if token["issuance_succeeds"]
-     *                        .as_bool()
-     *                        .expect("Could not check if issuance should succeed")
-     *                    {
-     *                        assert_ok!(Asset::create_token(
-     *                            Origin::signed(token_struct.owner),
-     *                            token_struct.name.clone(),
-     *                            token_struct.name.clone(),
-     *                            token_struct.total_supply,
-     *                            true
-     *                        ));
-     *
-     *                        // Also check that the new token matches what we asked to create
-     *                        assert_eq!(
-     *                            Asset::token_details(token_struct.name.clone()),
-     *                            token_struct
-     *                        );
-     *
-     *                        // Check that the issuer's balance corresponds to total supply
-     *                        assert_eq!(
-     *                            Asset::balance_of((token_struct.name, token_struct.owner)),
-     *                            token_struct.total_supply
-     *                        );
-     *
-     *                        // Add specified whitelist entries
-     *                        let whitelists = token["whitelist_entries"]
-     *                            .as_vec()
-     *                            .expect("Could not view token whitelist entries as vec");
-     *
-     *                        for wl_entry in whitelists {
-     *                            let investor = wl_entry["investor"]
-     *                                .as_str()
-     *                                .expect("Can't parse investor as string");
-     *                            let investor_id = accounts
-     *                                .get(&Yaml::String(investor.to_owned()))
-     *                                .expect("Can't get investor account record")["id"]
-     *                                .as_i64()
-     *                                .expect("Can't parse investor id as i64")
-     *                                as u64;
-     *
-     *                            let expiry = wl_entry["expiry"]
-     *                                .as_i64()
-     *                                .expect("Can't parse expiry as i64");
-     *
-     *                            let wl_id = wl_entry["whitelist_id"]
-     *                                .as_i64()
-     *                                .expect("Could not parse whitelist_id as i64")
-     *                                as u32;
-     *
-     *                            println!(
-     *                                "Token {}: processing whitelist entry for {}",
-     *                                ticker, investor
-     *                            );
-     *
-     *                            general_tm::Module::<Test>::add_to_whitelist(
-     *                                Origin::signed(owner_id),
-     *                                *ticker.into_bytes(),
-     *                                wl_id,
-     *                                investor_id,
-     *                                (now + Duration::hours(expiry)).timestamp() as u64,
-     *                            )
-     *                            .expect("Could not create whitelist entry");
-     *                        }
-     *                    } else {
-     *                        assert!(Asset::create_token(
-     *                            Origin::signed(token_struct.owner),
-     *                            token_struct.name.clone(),
-     *                            token_struct.name.clone(),
-     *                            token_struct.total_supply,
-     *                            true
-     *                        )
-     *                        .is_err());
-     *                    }
-     *                }
-     *
-     *                // Set up allowances
-     *                let allowances = case["allowances"]
-     *                    .as_vec()
-     *                    .expect("Could not view allowances as a vec");
-     *
-     *                for allowance in allowances {
-     *                    let sender = allowance["sender"]
-     *                        .as_str()
-     *                        .expect("Could not view sender as str");
-     *                    let sender_id = case["named_accounts"][sender]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view sender id as i64")
-     *                        as u64;
-     *                    let spender = allowance["spender"]
-     *                        .as_str()
-     *                        .expect("Could not view spender as str");
-     *                    let spender_id = case["named_accounts"][spender]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view sender id as i64")
-     *                        as u64;
-     *                    let amount = allowance["amount"]
-     *                        .as_i64()
-     *                        .expect("Could not view amount as i64")
-     *                        as u128;
-     *                    let ticker = allowance["ticker"]
-     *                        .as_str()
-     *                        .expect("Could not view ticker as str");
-     *                    let succeeds = allowance["succeeds"]
-     *                        .as_bool()
-     *                        .expect("Could not determine if allowance should succeed");
-     *
-     *                    if succeeds {
-     *                        assert_ok!(Asset::approve(
-     *                            Origin::signed(sender_id),
-     *                            *ticker.into_bytes(),
-     *                            spender_id,
-     *                            amount,
-     *                        ));
-     *                    } else {
-     *                        assert!(Asset::approve(
-     *                            Origin::signed(sender_id),
-     *                            *ticker.into_bytes(),
-     *                            spender_id,
-     *                            amount,
-     *                        )
-     *                        .is_err())
-     *                    }
-     *                }
-     *
-     *                println!("Transfers:");
-     *                // Perform regular transfers
-     *                let transfers = case["transfers"]
-     *                    .as_vec()
-     *                    .expect("Could not view transfers as vec");
-     *                for transfer in transfers {
-     *                    let from = transfer["from"]
-     *                        .as_str()
-     *                        .expect("Could not view from as str");
-     *                    let from_id = case["named_accounts"][from]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view from_id as i64")
-     *                        as u64;
-     *                    let to = transfer["to"].as_str().expect("Could not view to as str");
-     *                    let to_id = case["named_accounts"][to]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view to_id as i64")
-     *                        as u64;
-     *                    let amount = transfer["amount"]
-     *                        .as_i64()
-     *                        .expect("Could not view amount as i64")
-     *                        as u128;
-     *                    let ticker = transfer["ticker"]
-     *                        .as_str()
-     *                        .expect("Coule not view ticker as str")
-     *                        .to_owned();
-     *                    let succeeds = transfer["succeeds"]
-     *                        .as_bool()
-     *                        .expect("Could not view succeeds as bool");
-     *
-     *                    println!("{} of token {} from {} to {}", amount, ticker, from, to);
-     *                    let ticker = ticker.into_bytes();
-     *
-     *                    // Get sender's investor data
-     *                    let investor_data = <InvestorList<Test>>::get(from_id);
-     *
-     *                    println!("{}'s investor data: {:#?}", from, investor_data);
-     *
-     *                    if succeeds {
-     *                        assert_ok!(Asset::transfer(
-     *                            Origin::signed(from_id),
-     *                            ticker,
-     *                            to_id,
-     *                            amount
-     *                        ));
-     *                    } else {
-     *                        assert!(
-     *                            Asset::transfer(Origin::signed(from_id), ticker, to_id, amount)
-     *                                .is_err()
-     *                        );
-     *                    }
-     *                }
-     *
-     *                println!("Approval-based transfers:");
-     *                // Perform allowance transfers
-     *                let transfer_froms = case["transfer_froms"]
-     *                    .as_vec()
-     *                    .expect("Could not view transfer_froms as vec");
-     *                for transfer_from in transfer_froms {
-     *                    let from = transfer_from["from"]
-     *                        .as_str()
-     *                        .expect("Could not view from as str");
-     *                    let from_id = case["named_accounts"][from]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view from_id as i64")
-     *                        as u64;
-     *                    let spender = transfer_from["spender"]
-     *                        .as_str()
-     *                        .expect("Could not view spender as str");
-     *                    let spender_id = case["named_accounts"][spender]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view spender_id as i64")
-     *                        as u64;
-     *                    let to = transfer_from["to"]
-     *                        .as_str()
-     *                        .expect("Could not view to as str");
-     *                    let to_id = case["named_accounts"][to]["id"]
-     *                        .as_i64()
-     *                        .expect("Could not view to_id as i64")
-     *                        as u64;
-     *                    let amount = transfer_from["amount"]
-     *                        .as_i64()
-     *                        .expect("Could not view amount as i64")
-     *                        as u128;
-     *                    let ticker = transfer_from["ticker"]
-     *                        .as_str()
-     *                        .expect("Coule not view ticker as str")
-     *                        .to_owned();
-     *                    let succeeds = transfer_from["succeeds"]
-     *                        .as_bool()
-     *                        .expect("Could not view succeeds as bool");
-     *
-     *                    println!(
-     *                        "{} of token {} from {} to {} spent by {}",
-     *                        amount, ticker, from, to, spender
-     *                    );
-     *                    let ticker = ticker.into_bytes();
-     *
-     *                    // Get sender's investor data
-     *                    let investor_data = <InvestorList<Test>>::get(spender_id);
-     *
-     *                    println!("{}'s investor data: {:#?}", from, investor_data);
-     *
-     *                    if succeeds {
-     *                        assert_ok!(Asset::transfer_from(
-     *                            Origin::signed(spender_id),
-     *                            ticker,
-     *                            from_id,
-     *                            to_id,
-     *                            amount
-     *                        ));
-     *                    } else {
-     *                        assert!(Asset::transfer_from(
-     *                            Origin::signed(from_id),
-     *                            ticker,
-     *                            from_id,
-     *                            to_id,
-     *                            amount
-     *                        )
-     *                        .is_err());
-     *                    }
-     *                }
-     *            });
-     *        }
-     *    }
-     */
 }
