+++ runtime/src/test/mod.rs
@@ -1,5 +1,7 @@
 pub mod storage;
 pub use storage::TestStorage;
 
-mod identity;
+mod asset_test;
+mod identity_test;
 mod multisig;
+mod statistics_test;
